feat(activities): add logs
